+++ packages/core-typings/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocket.chat/core-typings
 
+## 6.3.3
+
+## 6.3.2
+
+## 6.3.1
+
 ## 6.3.0
 
 ### Minor Changes